Update Example11_DMP_SENSOR_LINEAR
This uses the onboard DMP rotation sensor to adjust the XYZ readings from the accelerometer into a constant orientation in the real world. Gravity is always down.

Due to drift in the Quat6 sensor, the gravity vector is periodically adjusted when the sensor is stationary. If we can get the Quat 9 reading to be more stable, then this step might be omitted.